+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.1
+
+- `RawSocketAsSocket`: finalize implementation.
+
+- New `RawServerSocketAsServerSocket`: `asSocket`
+- New `RawServerSocketExtension`: `asServerSocket`
+
+- `MultiDomainSecureServer`:
+  - Added `asServerSocket`.
+
 ## 1.0.0
 
 - Initial version.